+++ app/build.gradle
@@ -58,6 +58,8 @@ dependencies {
     implementation 'androidx.compose.material:material:1.1.1'
     implementation "androidx.navigation:navigation-compose:2.5.3"
 
+    implementation "com.google.accompanist:accompanist-navigation-animation:0.29.0-alpha"
+
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
