--- app/src/main/java/com/example/jetpackcomposeinstagram/navigation/AppNavigationRoutes.kt
@@ -1,5 +1,0 @@
-package com.example.jetpackcomposeinstagram.navigation
-
-sealed class routes(var routes: String){
-    object ListFruitItemsScreen: routes("listfruititemsscreen")
-}
