+++ app/src/main/java/com/example/jetpackcomposeinstagram/navigator/NavGraphNavigationRoutes.kt
@@ -1,0 +1,5 @@
+package com.example.jetpackcomposeinstagram.navigator
+
+sealed class Navgraphroutes(var routes: String){
+    object ListFruitItemsScreen: Navgraphroutes("listfruititemsscreen")
+}
